maslow/requirements.txt
Django>=4.2
python-dotenv
google-generativeai
pillow
django-crispy-forms
crispy-bootstrap5
